Create Longitudinal and latitudinal Pattern
First we are going to import the modules namely – Pandas, Matplotlib, and Numpy. Second is the import of the above mentioned csv file from the directory (change the path to the directory holding the csv file using >>>cd directory_address) into the variable birddata.
“bird_names = pd.unique(birddata.bird_name)” is used to find all the unique bird names from the csv file and save it to the variable bird_names using pandas dataframe and unique() function.
Now we will select the latitudinal and longitudinal data as mentioned in the code line 14-17 where we specify the variable ix to contain data of the column named “bird_name” from the csv file having name of the bird as “Eric”. 
We use the matplotlib function, figure() to initialize its size as 7 x 7 and plot it using the plot() function. The parameters inside the function plot() i.e. x, y and “b.” are specifying to use longitude data along x axis, latitude along y and b=blue, . = circles in the visualization.