Create Date and Time
We will now be including datetime library along with Pandas, Matplotlib, and Numpy.
The code line 7 – 9, describe creating an empty list “timestamps” and appending date-time data of the birds to it.
>>>datetime.datetime.today() #returns the current Date (yy-mm-dd) & time (h:m:s). 
>>>date_str[:-3] #slices/removes the UTC +00 coordinated time stamps.
>>>datetime.datetime.strptime (date_str[:-3], “%Y-%m-%d %H:%M:%S”) ,the time- stamp strings from date_str are converted to datetime object to be worked upon.
“%Y-%m-%d %H: %M: %S” is the Year-Month-Date and Hour-Minute-Second format”
The next step for us is to construct a panda series object and insert the timestamp from our Python list into that object. We can then append the panda series as a new column in my bird data, data frame. Line of code 15 -23, we create a list that captures the amount of time that has elapsed since the start of collection of data for the “Eric”.
Finally we plot the observation (reference points at constant distances) along x axis vs elapsed time (in days) along y axis. We label our plot using xlabel() and ylabel() as Observation and Elapsed time (days) respectively along x and y axis.